+++ pyqgisservercontrib/core/filters.py
@@ -11,7 +11,7 @@
 
     Declare filters as
 
-    @blockingfilter(pri=100, uri="/ows/")
+    @blockingfilter(uri="/ows/")
     def myfilter1( handller: tornado.web.RequestHandler, *args ) -> None:
         ...
 
@@ -28,53 +28,70 @@
     Modules implementing filters must register their filters using setup.py entry_points with the 
     key 'pyqgisserver_filters'
 
-    def register_filters():
-        # Do initialization stuff
-        return [ myfilter1, myfilter2 ]
+    policy_service.add_policy([myfilter1, myfilter2], pri=1000)
 """
 
 import logging
-import tornado.web
+import re
 
-from typing import Coroutine, Callable, Awaitable, Optional
+from tornado.httputil import HTTPServerRequest
+
+from typing import (
+    Union,
+    Tuple,
+    Callable, 
+    Awaitable, 
+    Optional,
+)
 
 LOGGER = logging.getLogger('SRVLOG')
 
 
-class ServerFilter:
-    pass
+class _FilterBase:
 
+    def __init__(self, match: Optional[Union[str,re.Pattern]]=None, repl: Optional[str]=None):
+        if isinstance(match, str):
+            match = re.compile(match, re.IGNORECASE)
+        self.pattern = match
+        self.repl = repl
+        self.match_args = []
+        self.match_kwargs = {}
 
-class asyncfilter(ServerFilter):
-    """ Decorator for asynchronous request filter
+    def __str__(self) -> str:
+        return f"_FilterBase<{hex(id(self))}>(match={self.pattern}, repl={self.repl})"
+    def match(self, path: str) -> Tuple[bool,str]:
+        """ Check uri against pattern
+        """
+        if self.pattern:
+            match = self.pattern.match(path)
+            if match:
+                # match.groups() includes both named and
+                # unnamed groups, we want to use either groups
+                # or groupdict but not both. 
+                if self.pattern.groupindex:
+                    self.match_kwargs = match.groupdict()
+                else:
+                    self.match_args = match.groups() 
+                if self.repl:
+                    path = self.pattern.sub(self.repl, path, count=1)
+                return True, path
+        else:
+            # Match everything
+            return True, path
 
-        Assume that the function return a  coroutine
-    """
+        return False, None
 
-    def __init__(self, pri=0, uri=""):
-        self.pri = pri
-        self.uri = uri
-
-    def __call__(self, fn: Callable[[tornado.web.RequestHandler], Awaitable[None]]) -> 'asyncfilter':
+    def __call__(self, fn: Callable):
         self.fn = fn
         return self
 
-    def apply( self, handler: tornado.web.RequestHandler ) -> Coroutine:
-        return self.fn(handler)
+    def apply(self, request: HTTPServerRequest) -> Optional[Awaitable]:
+        self.fn(request, *self.match_args, **self.match_kwargs)
+        return
 
 
-class blockingfilter(ServerFilter):
+class policy_filter(_FilterBase):
     """ Decorator for synchronous request filter
     """
-
-    def __init__(self, pri=0, uri=""):
-        self.pri = pri
-        self.uri = uri
-
-    def __call__(self, fn):
-        self.fn = fn
-        return self
-
-    async def apply( self, handler: tornado.web.RequestHandler ) -> Optional[str]:
-        return self.fn(handler)
+    pass