+++ tests/unittests/test_getcapabilities.py
@@ -31,7 +31,7 @@ class Tests(HTTPTestCase):
         """
         urlref = urlparse('https://my.proxy.loc:9999/anywhere/')
         rv = self.client.get("?MAP=france_parts.qgs&SERVICE=WMS&request=GetCapabilities", 
-                             headers={ 'X-Forwarded-Url': urlref.geturl() } )
+                             headers={ 'X-Forwarded-Url': urlref.geturl() })
 
         assert rv.status_code == 200
         assert rv.headers['Content-Type'] == 'text/xml; charset=utf-8'